refactor: :recycle: Botonera para la PomoPage refactorizada a un componente que maneja sólo lógica de configuraciones del timer
